backend/requirements.txt
# Core dependencies - compatible versions to avoid conflicts
pg8000==1.31.2
psycopg2==2.9.9
python-json-logger==2.0.7
alembic==1.13.2
SQLAlchemy==2.0.30
Flask==3.0.3
Flask-CORS==6.0.0
PyJWT==2.10.1
redis==5.0.4
bcrypt==4.1.3
boto3==1.28.57
psutil==5.9.8

# Schema validation
pydantic[email]>=2.6,<3

# Testing dependencies
pytest
pytest-cov
pytest-mock
pytest-rerunfailures
testcontainers

# Dev dependencies
flake8

# Security patches - urllib3 version limited by boto3 compatibility
# SECURITY NOTE: urllib3 1.26.20 has vulnerability GHSA-pq67-6m6q-mj2v (redirects)
# Cannot update to >=2.5.0 due to boto3 dependency constraint urllib3<1.27
# Risk mitigation: Application doesn't use urllib3 redirects directly
urllib3>=1.26.0,<2.0.0
